3D window rebuilt, better functionality and cleaner code
